gemini-chat/requirements.txt
chainlit
google-generativeai
python-dotenv